Update readme, remove unnecessary things in datasets.
