+++ README.md
@@ -25,11 +25,12 @@ python setup.py develop
 ```
 
 ### Data preparation
-**Step 0.** Download nuScenes official dataset.
+**Step 0.** Download nuScenes and waymo official dataset.
 
 **Step 1.** Symlink the dataset root to `./data/`.
 ```
 ln -s [nuscenes root] ./data/
+ln -s [waymo root] ./data/
 ```
 The directory will be as follows.
 ```
@@ -40,15 +41,21 @@ BEVDepth
 │   │   ├── samples
 │   │   ├── sweeps
 │   │   ├── v1.0-test
-|   |   ├── v1.0-trainval
+│   │   ├── v1.0-trainval
+│   ├── waymo
+│   │   ├── v1.4
+│   │   │   ├── training
+│   │   │   ├── validation
+│   │   │   ├── testing
 ```
 **Step 2.** Prepare infos.
+Prepare nuScenes data.
 ```
 python scripts/gen_info.py
 ```
-**Step 3.** Prepare depth gt.
+Prepare waymo data.
 ```
-python scripts/gen_depth_gt.py
+python scripts/gen_info.py waymo v1.4
 ```
 
 ### Tutorials
@@ -61,18 +68,18 @@ python [EXP_PATH] --amp_backend native -b 8 --gpus 8
 python [EXP_PATH] --ckpt_path [CKPT_PATH] -e -b 8 --gpus 8
 ```
 
-### Benchmark
+### Benchmark on nuScenes
 |Exp |EMA| CBGS |mAP |mATE| mASE | mAOE |mAVE| mAAE | NDS | weights |
 | ------ | :---: | :---: | :---:       |:---:     |:---:  | :---: | :----: | :----: | :----: | :----: |
-|[R50](exps/bev_depth_lss_r50_256x704_128x128_24e_2key_ema.py)| | |0.3304| 0.7021| 0.2795| 0.5346| 0.5530| 0.2274| 0.4355 | [github](https://github.com/Megvii-BaseDetection/BEVDepth/releases/download/v0.0.2/bev_depth_lss_r50_256x704_128x128_24e_2key.pth)
-|[R50](exps/bev_depth_lss_r50_256x704_128x128_24e_2key_ema.py)|√ | |0.3329 |  0.6832     |0.2761 | 0.5446 | 0.5258 | 0.2259 | 0.4409 | [github](https://github.com/Megvii-BaseDetection/BEVDepth/releases/download/v0.0.2/bev_depth_lss_r50_256x704_128x128_24e_2key_ema.pth)
-|[R50](exps/bev_depth_lss_r50_256x704_128x128_20e_cbgs_2key_da.py)| |√ |0.3484| 0.6159| 0.2716| 0.4144| 0.4402| 0.1954| 0.4805 | [github](https://github.com/Megvii-BaseDetection/BEVDepth/releases/download/v0.0.2/bev_depth_lss_r50_256x704_128x128_20e_cbgs_2key_da.pth)
-|[R50](exps/bev_depth_lss_r50_256x704_128x128_20e_cbgs_2key_da_ema.py)|√  |√ |0.3589 |  0.6119     |0.2692 | 0.5074 | 0.4086 | 0.2009 | 0.4797 | [github](https://github.com/Megvii-BaseDetection/BEVDepth/releases/download/v0.0.2/bev_depth_lss_r50_256x704_128x128_20e_cbgs_2key_da_ema.pth) |
+|[R50](exps/nuscenes/mv/bev_depth_lss_r50_256x704_128x128_24e_2key_ema.py)| | |0.3304| 0.7021| 0.2795| 0.5346| 0.5530| 0.2274| 0.4355 | [github](https://github.com/Megvii-BaseDetection/BEVDepth/releases/download/v0.0.2/bev_depth_lss_r50_256x704_128x128_24e_2key.pth)
+|[R50](exps/nuscenes/mv/bev_depth_lss_r50_256x704_128x128_24e_2key_ema.py)|√ | |0.3329 |  0.6832     |0.2761 | 0.5446 | 0.5258 | 0.2259 | 0.4409 | [github](https://github.com/Megvii-BaseDetection/BEVDepth/releases/download/v0.0.2/bev_depth_lss_r50_256x704_128x128_24e_2key_ema.pth)
+|[R50](exps/nuscenes/mv/bev_depth_lss_r50_256x704_128x128_20e_cbgs_2key_da.py)| |√ |0.3484| 0.6159| 0.2716| 0.4144| 0.4402| 0.1954| 0.4805 | [github](https://github.com/Megvii-BaseDetection/BEVDepth/releases/download/v0.0.2/bev_depth_lss_r50_256x704_128x128_20e_cbgs_2key_da.pth)
+|[R50](exps/nuscenes/mv/bev_depth_lss_r50_256x704_128x128_20e_cbgs_2key_da_ema.py)|√  |√ |0.3589 |  0.6119     |0.2692 | 0.5074 | 0.4086 | 0.2009 | 0.4797 | [github](https://github.com/Megvii-BaseDetection/BEVDepth/releases/download/v0.0.2/bev_depth_lss_r50_256x704_128x128_20e_cbgs_2key_da_ema.pth) |
 
 ## FAQ
 
 ### EMA
-- The results are differnt between evaluation during training and evaluation from ckpt.
+- The results are different between evaluation during training and evaluation from ckpt.
 
 Due to the working mechanism of EMA, the model parameters saved by ckpt are different from the model parameters used in the training stage.
